finish 99%

+++ database/migrations/2020_05_27_132152_create_persuratans_table.php
@@ -33,8 +33,9 @@ class CreatePersuratansTable extends Migration
             $table->date('tgl_masa_berlaku')->nullable();
             $table->enum('status_surat', ['Proses', 'Selesai']);
            
+            
             $table->integer('id_warga')->unsigned();
-            $table->bigInteger('id')->unsigned();
+            $table->integer('id')->unsigned();
             
             $table->timestamp('created_at')->useCurrent();
             $table->timestamp('updated_at')->useCurrent();
